splitted EncoderDecoder and getEncoding
